Fix issues in post-link scripts

+++ pgplot/post-link.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
 
 cd ${PREFIX}/lib
-ln -s libcpgplot.so libcpgplot.so.5
+if [ ! -e libcpgplot.so.5 ];
+then
+    ln -s libcpgplot.so libcpgplot.so.5
+fi
